Rename solo-cli to brkt-cli
Rename solo-cli to brkt-cli, since "Solo" is not an external product
name.  Fix the "brkt" launcher script so that it doesn't have
dependencies on my dev environment.
